fix: move span key to avoid dependancy
